requirements.txt
click>=8.3.1
fastapi>=0.128.0
loguru>=0.7.3
psutil>=7.2.1
pyexecjs>=1.5.1
pyperclip>=1.11.0
requests>=2.32.5
ujson>=5.11.0
uvicorn>=0.40.0

# Windows requirements
pywebview>=6.1; sys_platform != "linux"

# linux requirements
pywebview[gtk]>=6.1; sys_platform == "linux"

# 用户认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Optional dependencies for comment analysis features
# 浏览器自动化爬取（备用方案）
DrissionPage>=4.0.0; python_version >= "3.8"

# 数据分析可视化
jieba>=0.42.1
wordcloud>=1.9.0
matplotlib>=3.5.0
pyecharts>=2.0.0

# 可选：高级分析
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# MySQL 数据库支持
pymysql>=1.1.0
cryptography>=41.0.0